+++ PSNovusTools/Private/Invoke-SSHSessionItemBash.ps1
@@ -1,0 +1,9 @@
+function Invoke-SSHSessionItemBash([SSH.SshSession]$session, [string]$plainTextPassword,[string]$sshBashCommand) {
+
+    $sshCommand = "echo '$plainTextPassword' | sudo -S $sshBashCommand"
+    
+    $sshresult = Invoke-SSHSessionItem -session $session -sshCommand $sshCommand
+ 
+
+    return $sshresult
+}