+++ PSNovusTools/Private/Invoke-SSHSessionItemPowershell.ps1
@@ -1,0 +1,11 @@
+function Invoke-SSHSessionItemPowershell([SSH.SshSession]$session, [string]$sshPowershellCommand) {
+
+    $sshCommand = @"
+powershell -Command "$sshPowershellCommand"
+"@
+    
+    $sshresult = Invoke-SSHSessionItem -session $session -sshCommand $sshCommand
+ 
+
+    return $sshresult
+}